feat: add runtime detector weight tuning

+++ analysis/src/c2hunter_analysis/domain.py
@@ -228,7 +228,7 @@ class Evidence:
 @dataclass(frozen=True)
 class ScoreAdjustment:
     kind: str
-    points: int
+    points: float
     explanation: str
